policy

+++ app/Http/Controllers/JurusanController.php
@@ -28,6 +28,8 @@ class JurusanController extends Controller
      */
     public function store(Request $request)
     {
+        // batasi hak akses untuk proses store
+        $this->authorize('create', Jurusan::class);
         $validateData = $request->validate([
             'nama_jurusan' => 'required',
             'nama_dekan' => 'required',
@@ -76,6 +78,7 @@ class JurusanController extends Controller
      */
     public function destroy(Jurusan $jurusan)
     {
+        $this->authorize('delete',$jurusan);
         $jurusan->delete();
         return redirect('/')
             ->with('pesan', "Jurusan $jurusan->nama_jurusan berhasil dihapus");
